# docs/openapi-v0.4.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Camera Market Strategy System",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/strategies": {
      "get": {
        "tags": [
          "strategies"
        ],
        "summary": "List Strategies",
        "operationId": "list_strategies_api_strategies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StrategyOut"
                  },
                  "type": "array",
                  "title": "Response List Strategies Api Strategies Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "strategies"
        ],
        "summary": "Create Strategy",
        "operationId": "create_strategy_api_strategies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/strategies/{strategy_id}": {
      "put": {
        "tags": [
          "strategies"
        ],
        "summary": "Update Strategy",
        "operationId": "update_strategy_api_strategies__strategy_id__put",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Strategy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/strategies/{strategy_id}/copy": {
      "post": {
        "tags": [
          "strategies"
        ],
        "summary": "Copy Strategy",
        "operationId": "copy_strategy_api_strategies__strategy_id__copy_post",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Strategy Id"
            }
          },
          {
            "name": "user_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ronin",
              "title": "User Name"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/signals": {
      "get": {
        "tags": [
          "signals"
        ],
        "summary": "List Signals",
        "operationId": "list_signals_api_signals_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "current_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Current Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SignalOut"
                  },
                  "title": "Response List Signals Api Signals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/signals/today": {
      "get": {
        "tags": [
          "signals"
        ],
        "summary": "Today Signals",
        "operationId": "today_signals_api_signals_today_get",
        "parameters": [
          {
            "name": "current_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Current Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SignalOut"
                  },
                  "title": "Response Today Signals Api Signals Today Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/signals/product/{product_id}": {
      "get": {
        "tags": [
          "signals"
        ],
        "summary": "Product Signals",
        "operationId": "product_signals_api_signals_product__product_id__get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          },
          {
            "name": "current_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Current Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SignalOut"
                  },
                  "title": "Response Product Signals Api Signals Product  Product Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reports/generate": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Generate Report",
        "operationId": "generate_report_api_reports_generate_post",
        "parameters": [
          {
            "name": "report_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Report Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/{report_id}/download": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Download Report",
        "operationId": "download_report_api_reports__report_id__download_get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/quant/products/{product_id}/indicators": {
      "get": {
        "tags": [
          "quant"
        ],
        "summary": "Indicators",
        "operationId": "indicators_api_quant_products__product_id__indicators_get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          },
          {
            "name": "window_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3650,
              "minimum": 7,
              "default": 180,
              "title": "Window Days"
            }
          },
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 12,
              "default": "CNY",
              "title": "Currency"
            }
          },
          {
            "name": "include_visible",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Visible"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuantIndicatorsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/quant/backtests": {
      "post": {
        "tags": [
          "quant"
        ],
        "summary": "Backtest",
        "operationId": "backtest_api_quant_backtests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BacktestRequest": {
        "properties": {
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "strategy_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Id"
          },
          "trigger_price": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Price"
          },
          "strong_buy_price": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Strong Buy Price"
          },
          "window_days": {
            "type": "integer",
            "maximum": 3650.0,
            "minimum": 7.0,
            "title": "Window Days",
            "default": 180
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "CNY"
          },
          "include_visible_prices": {
            "type": "boolean",
            "title": "Include Visible Prices",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "product_id"
        ],
        "title": "BacktestRequest"
      },
...
      "DailyReportOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "report_date": {
            "type": "string",
            "format": "date",
            "title": "Report Date"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "markdown_content": {
            "type": "string",
            "title": "Markdown Content"
          },
          "chart_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chart Path"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "report_date",
          "title",
          "summary",
          "markdown_content",
          "chart_path",
          "created_at"
        ],
        "title": "DailyReportOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SignalOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "strategy_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Id"
          },
          "price_record_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Record Id"
          },
          "signal_type": {
            "type": "string",
            "title": "Signal Type"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "triggered": {
            "type": "boolean",
            "title": "Triggered"
          },
          "is_current": {
            "type": "boolean",
            "title": "Is Current",
            "default": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "product_id",
          "strategy_id",
          "price_record_id",
          "signal_type",
          "message",
          "triggered",
          "created_at"
        ],
        "title": "SignalOut"
      },
      "StrategyCreate": {
        "properties": {
          "user_name": {
            "type": "string",
            "title": "User Name",
            "default": "ronin"
          },
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "strategy_name": {
            "type": "string",
            "title": "Strategy Name"
          },
          "trigger_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Price"
          },
          "strong_buy_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strong Buy Price"
          },
          "watch_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Watch Price"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "CNY"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "value_hunter"
          },
          "max_price_age_hours": {
            "type": "integer",
            "maximum": 720.0,
            "minimum": 1.0,
            "title": "Max Price Age Hours",
            "default": 24
          },
          "near_target_pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Near Target Pct",
            "default": 5.0
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "strategy_name"
        ],
        "title": "StrategyCreate"
      },
      "StrategyOut": {
        "properties": {
          "user_name": {
            "type": "string",
            "title": "User Name",
            "default": "ronin"
          },
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "strategy_name": {
            "type": "string",
            "title": "Strategy Name"
          },
          "trigger_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Price"
          },
          "strong_buy_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strong Buy Price"
          },
          "watch_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Watch Price"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "CNY"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "value_hunter"
          },
          "max_price_age_hours": {
            "type": "integer",
            "maximum": 720.0,
            "minimum": 1.0,
            "title": "Max Price Age Hours",
            "default": 24
          },
          "near_target_pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Near Target Pct",
            "default": 5.0
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "strategy_name",
          "id",
          "created_at"
        ],
        "title": "StrategyOut"
      },
      "StrategyUpdate": {
        "properties": {
          "strategy_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Name"
          },
          "trigger_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Price"
          },
          "strong_buy_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strong Buy Price"
          },
          "watch_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Watch Price"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "max_price_age_hours": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 720.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Price Age Hours"
          },
          "near_target_pct": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Near Target Pct"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "StrategyUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}